Use our own logger implementation, after much struggle with ocaml Format.
I dislike the Lwt_log for a few reasons:

When it logs to stderr (the default logger), it is run as job, so it
runs the write(2) in another system thread, which makes me feel dirty.

I hate the idea of desynchronized logs, logs should be synchronized,
therefore the should block. On my previous attempt, I was pretty much
Lwt.bind`ing every log call.

Thanks to @nojb for helping me greatly with the mess that ocaml format
strings are.