Add the humble begginings of a dhcpd ISC configuration parser using menhir.
The lexer is implemented in lexer.mll and the parser in parser.mly.

config_parser is the module where the ball starts rolling, sadly I had to add
almost empty config.ml and config.mli to break module interdependancy as merlin
cannot export things in its .mli.

For now this is enough to parse my own dhcpd.conf.
Although I'm aiming to be compatible with ISC's dhcpd.conf, I don't think full
compatibility is desired, I seriously doubt implementing "group" support is any
worth, but that might change.

This took me a lot of work, special thanks to various at #ocaml for the
assistance with menhir.