Add ctypes.foreign and if_nametoindex, if_indextoname.
Uppon receiving a packet, we need to know which interface it came from. Since
the packet is not ip, and most likely sent to the broadcast mac address, we
can't use the destination address to figure out which interface it was delivered
to. We need to know which interface it was delivered to in order to be able to
match the subnet with the client.

The only feasible way seems to be using control ancilliary data with recv_msg
and matching the interface index number, but we had no way to know which
interface name had which index number, so we add if_nametoindex.