Implement recv_if, a read(2) which returns the received interface.
Snatched this from my ongoing work on ExtUnix, it also adds lwt and ctypes
wrappers.

Only implemented for BSD so far (IP_RECVIF), next step is support PKTINFO.
I've used C stubs instead of ctypes since it is not clear to me yet how to do
the later.

I guess it makes sense to move this to a proper library someday. Will also have
to sort it out how to make this compatible with mirage in the future.

Also add reqif() to call the appropriate setsockopt.