Allow an expired lease to be re-used in INIT-REBOOT and RENEWING/REBINDING.
When a lease had expired, we were NAKing the address, this is stupid, yet
permited by the standard. The wording on the RFC says:

"""
If the network is correct, then the DHCP server should check if
the client's notion of its IP address is correct.
"""

So if a lease has expired, the notion of its address is still correct as long as
the address wasn't claimed by someone else, in this case we should renew the
lease.