Big changes, make config functorize over INTERFACE, add S.mli.
Add a S.mli for the module signatures, moving most of the code from
dhcp_server.mli.

Add a SERVER.parse_networks, this is needed so that charrua-unix can upfront
know all the used networks by the configuration file, it then limits the
interfaces it uses by theses networks. If that is not the case, charrua-unix
would have to open a bunch of Rawlinks that would go unused.

Cleaned up some code in config.ml.

Renamed Dhcp_server.start to Dhcp_server.create.