Implement the fixed address statement.
We fake leases, when we create the lease database, we also create a secondary
fixed address table of Macaddr.t -> Ipaddr.V4.t.

When a lookup for a fixed map happens, we create a lease on the fly. When an
insert of that lease happens, we ignore. This way the server logic can be
ignorant if it is dealing with a real lease or a fake one.