Implement -D, daemonize option.
Lwt is a bit painful, I don't like the log infrastructure so I'll keep using my
own. Lwt_daemon.daemonize redirects the output of stdout and stderr to syslog,
which is great, especially for exceptions.

There are basic two problems with the syslog implementation:
1 - Instead of using syslog(3), it opens the socket manually, so it needs access
to /var/bla/sock.
2 - There is no openlog.

I had to manually add the paths for OS X, making 1 not so seriously.

But I would need an openlog, since it needs to happen _before_ we call
go_safe(). At least, it opens and keeps the socket open on the first log call,
that's why we just print "daemonized" at some point, if you remove that, boom,
openlog will happen after dropping privileges and there won't be any cake.