Merge back Dhcp_wire into Dhcp.
I'm still unsure if we should keep Dhcp_wire or Dhcp, keeping both does not
please me at all.

Also remove Dhcp.mli, until this is finished, keeping a .mli is a pain in the
ass as the interface keeps changing. Once this is ready, we recreate a proper
and documented mli.