Major rewrite, pull out IO from Dhcp_server.
This removes the Interface abstraction as well as all the functors from the
code.

Dhcp_server is broken down into two sub-modules:

Config - deals with the read-only server configuration.
Input - Server logic and main runtime, driven from an input function.

Input is almost pure (Leases are still not pure), it doesn't do any IO, you give
Input.input_pkt a Dhcp_wire.pkt and you get a Input.result, which at this point
can be an Error, a Warning, a Reply pkt, or Silence (nothing).

This way the caller is resposible for sending the reply and/or logging the
event, I've decided to change this after the Functors brought much complication
to the code and Hannes Mehnert suggested returning actions from the server logic.

The server code does not depend on Lwt anymore, it only deals with the logic
itself. Dhcp_logger is gone as well. Config.ml has been moved to
Dhcp_server.Config.

I've tried to broken this into smaller commits but got pissed at some point and
squashed them all.