Finish Dhcp_server.collect_replies.
Now we're able to match every parameter request with a possible dhcp_option
reply.

SUBNET_MASK is still being sent even when not requested and we ignore it in
collect_replies, this should be changed, but it requires the configuration
builder to insert the dhcp option and I can't test such change now.