Pass time as seconds in Int32.t.
This requires now a monotonic clock as Int32.t is not good enough for storing
walltime. 32bits give us up to 2038, but since this is a signed int, it's way
less, we could hold it until 2038 but then we have to be ueber careful when
comparing signed ints as uints.