add renewal logic; expose Dhcp_client_lwt
Previously, the only exposed modules were a pure Dhcp_client and a
module for use in Mirage, implementing the client timing logic and
exposing only an ipv4_config.

This changeset exposes a Dhcp_client_lwt module which
handles timing logic and exposes the full lease information once the
lease is obtained, for use in a wider variety of contexts (e.g., a
client which also sets network parameters according to the lease in a
traditional OS).

Since the client may now be used in a context where proper renewal is
possible, add logic and tests for handling it properly.
As part of this change, Dhcp_client.input now returns a polyvar
rather than an option, indicating whether the message has a new lease.