make the default timeout 4s, not 3s
On my local Xen system, the dnsmasq process running on the other side of
openvswitch often takes just short of 3s to send the first DHCPOFFER.
This causes the client to give up right before a successful lease
transaction and makes for some confusing log output.  Use a slightly
more conservative timeout (in the middle of the range allowed in the
RFC, rather than at the short end as before).