Fix dhcp client id
Clients which uses a Client Id with type 0xff could not get an
dhcp lease.
The Problem is in the Client Id type of 0xff. Charrua ignored the type
and set it internal to 0x00. When the client sent the "Request",
charrua could not find the matching Offer because of the mismatch in the client
id.

client | server
-> DISCOVER
<- OFFER
-> REQUEST
<- NAK!

Fixes #84

Note to previous self:
 - I've no idea how I fixed it. Can you explain it?