mirage client: use Dhcp_client_mirage module in Dhcp_ipv4 directly
There's no need for abstraction (over a module type) since we did not
develop any other DHCP implementation in recent years. Abstraction currently
comes with a price, hardcoded code in the mirage utility - which we want
to minimize.

This also removes the dependency on Mirage_protocols.ipv4_config.

It will require a version bump to avoid an old mirage utility to use this
new code (which module signature is different now).