Rewrite server lease database and fix a bunch of bugs.
The double reference between old Id_map and Addr_map is too confusing and prone
to mistakes, this rewrites so that Addr_map now only gives a Client_id and
Lease_map gives a Lease.t.

Instead of trying to recycle old leases on the fly and introducing unecessary
complications, just remove it all over. Code gets considerably simpler and the
caller is responsible for calling Lease.garbage_collect once in a while.

Expired leases are kept even if expire, if a client does a discover later on, it
might re-active/renew the old lease, this is desirable since he gets to have his
same old ip back.

Fix a bug where the server *MUST NOT* ack a lease during INIT-REBOOT, it should
remain silent, it should ack only when SELECTING or RENEW/REBIDING. With this
every ack becomes a lease renew, so we don't need the extra logic of ack ~renew.