Initial Eio bindings for charrua-unix
This adds a new binary `charruad-eio` (`charrua-unix-eio`), it does the same thing
as charrua-unix but with an Eio backend instead of the default Lwt.

There are two main differences from the Lwt version:

  - There is no syslog support on Eio at this time, so I've imported Logs-syslog
and use the synchronous UNIX backend.

 - There is also no `daemonize` function in Eio like Lwt was providing, so I've
written my own, this is not as nice as before since with a better syslog support
we can redirect stderr to syslog, so we at least get exceptions and whatnot in
the daemon.

This has been powering my home for the last 4 days, my wife hasn't killed me,
which makes me think it's working well.

While here, also adjust `charrua-unix` to rely on newer `rawlink-2.1`. Rawlink
has been split into
  - `rawlink`: good ol' blocking
  - `rawlink-lwt`: Lwt backend, used by `charrua-unix`
  - `rawlink-eio`: Eio new backend, used by `charrua-unix-eio`