Remove ppx_cstruct from dependency cone
The goal is to move to (a) cstruct-free and (b) smaller binary size.

This is the path towards using the same approach as ocaml-dns (a GADT map), and
also what ocaml-tls did in recent years (remove the not supported enums, etc.).
While this commit adds code, it removes ppx_cstruct. In subsequent commits,
we'll remove more code, and likely result in a smaller (and thus more manageable)
codebase.